[FEAT] add binding fragment
