+++ app/src/main/java/org/sopt/sample/presentation/ui/main/gallery/GalleryFragment.kt
@@ -6,15 +6,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import org.sopt.sample.R
+import org.sopt.sample.databinding.FragmentGalleryBinding
+import org.sopt.sample.presentation.common.base.BindingFragment
 
-class GalleryFragment : Fragment() {
+class GalleryFragment : BindingFragment<FragmentGalleryBinding>(R.layout.fragment_gallery) {
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_gallery, container, false)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
     }
 
 }
